use new header and rename display panel
