+++ client/apps/edit/components/display/components/magazine.jsx
