+++ client/apps/edit/components/display/components/partner.jsx
