+++ client/apps/edit/components/display/components/preview/email_preview.jsx
