+++ client/apps/edit/components/display/components/preview/magazine_preview.jsx
