--- client/apps/edit/components/display/form.jade
@@ -1,7 +1,0 @@
-- image = image || article.get('thumbnail_image')
-- context = context
-label
-  span.edit-required(
-    data-hidden=image ? 'true' : 'false'
-  ) #{context} Image
-  .edit-display__image-upload
