--- client/apps/edit/components/display/index.jade
@@ -1,152 +1,0 @@
-if sd.CURRENT_CHANNEL.get('type') === 'partner'
-  section#edit-display.edit-display--partner
-    header
-      h1 Prepare your article for Artsy and social media
-      h2 Choose a thumbnail image and title that captures the user&rsquo;s attention
-      img( src='/images/edit_prepare_example.png')
-    form.edit-display.edit-display--partner-inputs
-      .edit-display--partner-left
-        include form
-      .edit-display--partner-right
-        label.edit-display__headline
-          span.edit-required Title
-          textarea.bordered-input.edit-title-textarea.edit-display__textarea(
-            placeholder="Type a title that captures your reader's attention"
-            data-limit=97
-          )= article.get('thumbnail_title')
-          .edit-display__use-article-title Use article title
-          .edit-char-count-message Remaining Characters:
-            .edit-char-count
-else
-  section#edit-display.max-width-container
-    form.edit-display.admin-form-container
-
-      section.edit-display--magazine
-        .admin-form-section.edit-display__section-title
-          h1 Magazine
-          include ../../public/icons/edit_caret_down.svg
-        .edit-display__inputs
-          .admin-form-left
-            - context = 'Magazine'
-            include form
-            label.edit-display__headline
-              .label
-                span.title Magazine Headline
-                .edit-char-count
-              textarea.bordered-input.edit-title-textarea.edit-display__textarea(
-                data-limit=97
-                data-value='thumbnail_title'
-              )=article.get('thumbnail_title')
-              .edit-display__use-article-title Use Article Title
-            label.edit-display__description
-              .label
-                span.title Magazine Description
-                .edit-char-count
-              textarea.bordered-input.edit-display__textarea(
-                data-limit=160
-                data-value='description'
-              )=article.get('description')
-          .admin-form-right
-            .edit-display__preview
-              include templates/magazine_preview
-
-      //- Search
-      section.edit-display--social
-        .admin-form-section.edit-display__section-title
-          h1 Social
-          include ../../public/icons/edit_caret_down.svg
-        .edit-display__inputs
-          .admin-form-left
-            - image = article.get('social_thumbnail')
-            - context = 'Social'
-            include form
-            label.edit-display__headline
-              .label
-                span.title Social Headline
-                .edit-char-count
-              textarea.bordered-input.edit-display__textarea(
-                data-limit=97
-                data-value='social_title'
-              )=article.get('social_title')
-            label.edit-display__description
-              .label
-                span.title Social Description
-                .edit-char-count
-              textarea.bordered-input.edit-display__textarea(
-                data-limit=160
-                data-value='social_description'
-              )=article.get('social_description')
-          .admin-form-right
-            .edit-display__preview
-              include templates/social_preview
-
-      //- Search
-      section.edit-display--search
-        .admin-form-section.edit-display__section-title
-          h1 Search
-          include ../../public/icons/edit_caret_down.svg
-        .edit-display__inputs
-          .admin-form-left
-            label.edit-display__headline
-              .label
-                span.title Search Headline
-                .edit-char-count
-              textarea.bordered-input.edit-display__textarea(
-                data-limit=97
-                data-value='search_title'
-              )=article.get('search_title')
-            label.edit-display__description
-              .label
-                span.title Search Description
-                .edit-char-count
-              textarea.bordered-input.edit-display__textarea(
-                data-limit=160
-                data-value='search_description'
-              )=article.get('search_description')
-          .admin-form-right
-            .edit-display__preview
-              include templates/search_preview
-
-      //- Email
-      section.edit-display--email
-        .admin-form-section.edit-display__section-title
-          h1 Email
-          include ../../public/icons/edit_caret_down.svg
-        .edit-display__inputs
-          .admin-form-left
-            label.edit-email-image
-              span(
-                data-hidden=(article.get('email_metadata') && article.get('email_metadata').image_url) ? 'true' : 'false'
-              ) Email Thumbnail
-              .edit-display__image-upload
-            label.edit-display__headline
-              span.label Email Headline
-              textarea.bordered-input.edit-display__textarea(
-                name='headline'
-                data-value='headline'
-              )= article.getObjectAttribute('email_metadata','headline')
-            label Custom Text
-              input.bordered-input( name='custom_text' value="#{article.getObjectAttribute('email_metadata','custom_text')}")
-
-          .admin-form-right
-            .edit-display__preview
-              include templates/email_preview
-
-          .admin-form-left
-            label Author
-              input.bordered-input.edit-display__author(
-                name='author'
-                value="#{article.getObjectAttribute('email_metadata','author')}"
-                data-value='author'
-              )
-          .admin-form-right.edit-display__email-credits
-            .admin-form-left
-              label Credit Line
-                input.bordered-input( name='credit_line' value="#{article.getObjectAttribute('email_metadata','credit_line')}")
-            .admin-form-right
-              label Credit URL
-                input.bordered-input( name='credit_url' value="#{article.getObjectAttribute('email_metadata','credit_url')}")
-          label.admin-send-body-to-sailthru
-            .flat-checkbox
-              input( type='checkbox' name='send_body' value=1 checked=article.get('send_body') === true)
-            span Send Article Body To Sailthru
