--- client/apps/edit/components/display/templates/magazine_preview.jade
@@ -1,15 +1,0 @@
-.edit-display__prev-mag
-  .edit-display__prev-mag--left
-    .edit-display__prev-mag--date= article.date('published_at').format('MMMM Do')
-    .edit-display__prev-mag--headline= article.get('thumbnail_title')
-    .edit-display__prev-mag--byline= "By " + article.getByline()
-    .edit-display__prev-mag--social
-      include ../icons/social_group.svg
-  .edit-display__prev-mag--right
-    .edit-display__prev-mag--placeholder
-    .edit-display__prev-mag--placeholder
-      if article.get('thumbnail_image')
-        img(src=crop(article.get('thumbnail_image'),{width: 300, height: 200}))
-      else
-        .edit-display__prev--x
-    .edit-display__prev-mag--placeholder
