+++ client/apps/edit/components/display/test/magazine.test.js
