--- client/apps/edit/components/display2/index.styl
@@ -1,53 +1,0 @@
-.EditDisplay
-  padding-top 40px
-  .DropDownItem__content
-    margin-top 40px
-  .field-group.flat-checkbox
-    display flex
-    align-items center
-  .edit-display__preview
-    margin-left 20px
-  .title-fill
-    color gray-dark-color
-    garamond l-caption
-    text-align right
-    width 100%
-    margin-top 5px
-    cursor pointer
-
-.DisplayEmail
-  .col--right
-    display flex
-    flex-direction column
-    justify-content space-between
-
-  .field-group--inline
-    margin-left 20px
-    .field-group
-      width 100%
-      &:first-child
-        margin-right 15px
-      &:last-child
-        margin-left 15px
-
-.DisplayPartner
-  &__preview
-    background-color gray-lightest-color
-    text-align center
-    .header
-      padding 40px
-    img
-      width 100%
-    h1
-      garamond headline
-  &__edit
-    display flex
-    margin-top 40px
-  .CharacterLimit[data-type=textarea] .DraftEditor-root
-    min-height 166px
-    garamond headline
-
-#edit-display.edit-display--partner
-  margin-top 55px
-  .EditDisplay
-    padding-top 0px
