--- client/apps/edit/components/display2/preview.styl
@@ -1,185 +1,0 @@
-.EditDisplay
-
-  .edit-display__preview
-    border 1px solid gray-lighter-color
-    margin 20px 0
-
-  .edit-display__prev-mag
-    display flex
-    line-height 1.25
-
-    &--left
-      margin 0 35px 0 20px
-      display flex
-      flex-direction column
-      justify-content center
-    &--date
-      garamond()
-      line-height 1.5
-      font-size 9px
-      color gray-dark-color
-    &--byline
-      avant-garde()
-      font-size 9px
-      margin-bottom
-    &--headline
-      margin 8px 0
-    &--social
-      width auto
-      svg
-        width 10px
-        margin-right: 10px
-
-    &--right
-      margin-right 20px
-      display flex
-      flex-direction column
-    &--placeholder
-      width 300px
-      height 55px
-      margin-bottom 30px
-      background-color gray-lighter-color
-    &--placeholder:nth-child(2)
-      height 200px
-      display flex
-      justify-content center
-      align-items center
-    &--placeholder:last-child
-      margin-bottom 0px
-
-  .edit-display__prev-social
-    display flex
-    flex-direction column
-    background-color gray-lighter-color
-    align-items center
-    justify-content space-between
-    &--post
-      width 320px
-      min-height 34px
-      background-color white
-      margin-bottom 10px
-      padding 10px
-    &--post:last-child
-      margin-bottom 0px
-    &--content
-      border 1px solid gray-lighter-color
-    &--placeholder
-      // Facebook image dimensions
-      width 300px
-      height 157px
-      background-color gray-lighter-color
-      display flex
-      justify-content center
-      align-items center
-    &--meta
-      padding 5px
-      font-family 'Lucida Grande', 'Tahoma', 'Helvetica', 'Arial', 'sans-serif'
-    &--description
-      font-size 10px
-      line-height 12px
-    &--headline
-      font-family 'Times New Roman', 'serif'
-      font-size 16px
-      line-height 1.25
-    &--author
-      font-size 6.3px
-      text-transform uppercase
-      color gray-color
-    &--identity
-      display flex
-      padding-bottom 10px
-      svg
-        width 25px
-        height 25px
-        fill-color gray-color
-        margin-right 10px
-    &--icontainer
-      display flex
-      flex-direction column
-      justify-content space-between
-    &--name, &--hours-ago
-      font-family 'Lucida Grande', 'Tahoma', 'Helvetica', 'Arial', 'sans-serif'
-      font-size 10px
-      line-height 12px
-      color gray-color
-    &--hours-ago
-      font-size 8px
-
-  .edit-display__prev-search
-    padding 20px 100px
-    font-family 'Arial', 'sans-serif'
-    font-size 9.5px
-    line-height 12px
-    color black
-    &--result-headline
-      height 1.5em
-      width 60%
-      background-color gray-lighter-color
-      margin-bottom 3px
-    &--result-line
-      height 1em
-      background-color gray-lighter-color
-      margin-bottom 3px
-      &:nth-child(1)
-        width 70%
-      &:nth-child(2)
-        width 80%
-      &:last-child
-        width 95%
-        margin-bottom 15px
-    &--headline
-      font-size 12px
-      line-height 15px
-      color #1a0dab
-    &--slug
-      color #006621
-    &--description
-      display inline-block
-    &--date
-      display inline-block
-      color gray-dark-color
-
-  .edit-display__prev-email
-    text-align center
-    display flex
-    flex-direction column
-    justify-content center
-    align-items center
-    &--placeholder
-      width 320px
-      height 188px
-      margin-top 28px
-      background-color gray-lighter-color
-      display flex
-      justify-content center
-      align-items center
-    &--placeholder:last-child
-      height 30px
-    &--headline
-      garamond('l-body')
-      margin-top 15px
-      max-width 320px
-    &--author
-      font-family 'Arial', 'sans-serif'
-      text-transform uppercase
-      font-size 8px
-      font-color gray-color
-
-  .edit-display__prev
-    &--x
-      position relative
-      width 100px
-      height 100px
-      display flex
-      flex-direction column
-      justify-content center
-    &--x::before, &--x::after
-      position absolute
-      content ''
-      width 100%
-      height 1px
-      background-color white
-    &--x::before
-      transform rotate(45deg)
-    &--x::after
-      transform rotate(-45deg)
