+++ client/apps/edit/components/edit_container.jsx
@@ -5,7 +5,7 @@ import { connect } from 'react-redux'
 import * as Actions from 'client/actions/editActions'
 
 import { EditContent } from './content/index.jsx'
-import { EditDisplay } from './display2/index.jsx'
+import { EditDisplay } from './display/index.jsx'
 import { EditHeader } from './header/index.jsx'
 import EditAdmin from './admin/index'
