+++ client/apps/edit/components/header/index.jsx
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 import Icon from '@artsy/reaction-force/dist/Components/Icon'
-import { data as sd } from 'sharify'
+import colors from '@artsy/reaction-force/dist/Assets/Colors'
 
 export class EditHeader extends Component {
   static propTypes = {
@@ -13,11 +13,13 @@ export class EditHeader extends Component {
   }
 
   toggleSection = (activeSection) => {
+    // TODO - connect component to use redux actions directly
     this.props.actions.changeSection(activeSection)
   }
 
   isPublishable = () => {
-    return this.contentIsComplete() && this.displayIsComplete()
+    const { article } = this.props
+    return article.finishedContent() && article.finishedThumbnail()
   }
 
   onPublish = () => {
@@ -43,11 +45,11 @@ export class EditHeader extends Component {
     const { isSaving, isSaved } = this.props.edit
 
     if (isSaving) {
-      return 'limegreen'
+      return colors.greenRegular
     } else if (isSaved) {
       return 'black'
     } else {
-      return '#f7625a'
+      return colors.redMedium
     }
   }
 
@@ -80,19 +82,10 @@ export class EditHeader extends Component {
     }
   }
 
-  contentIsComplete = () => {
-    const { article } = this.props
-    return article.get('title') && article.get('title').length
-  }
-
-  displayIsComplete = () => {
-    const { article } = this.props
-    return article.get('thumbnail_title') && article.get('thumbnail_image')
-  }
-
   render () {
     const { article, actions, channel, user } = this.props
     const { activeSection, isDeleting } = this.props.edit
+    const { grayMedium, greenRegular } = colors
 
     return (
       <div className='EditHeader'>
@@ -108,7 +101,7 @@ export class EditHeader extends Component {
               <Icon
                 className='icon'
                 name='check'
-                color={this.contentIsComplete() ? 'limegreen' : '#ccc'}
+                color={article.finishedContent() ? greenRegular : grayMedium}
               />
             </button>
 
@@ -121,7 +114,7 @@ export class EditHeader extends Component {
               <Icon
                 className='icon'
                 name='check'
-                color={this.displayIsComplete() ? 'limegreen' : '#ccc'}
+                color={article.finishedThumbnail() ? greenRegular : grayMedium}
               />
             </button>
 
@@ -157,8 +150,7 @@ export class EditHeader extends Component {
 
         <div className='EditHeader__right'>
           <button
-            className='avant-garde-button'
-            style={{border: 'none'}}
+            className='avant-garde-button delete'
             onClick={this.onDelete}
           >
             {isDeleting ? 'Deleting...' : 'Delete'}
@@ -174,7 +166,7 @@ export class EditHeader extends Component {
             {this.getSaveText()}
           </button>
 
-          <a href={`${sd.FORCE_URL}/article/${article.get('slug')}`} target='_blank'>
+          <a href={article.getFullSlug()} target='_blank'>
             <button className='avant-garde-button'>
               {article.get('published') ? 'View' : 'Preview'}
             </button>
