+++ client/apps/edit/components/header/index.styl
@@ -20,6 +20,8 @@
     &[data-disabled=false],
     &[data-active=true]
       color black
+    &.delete
+      border none
 
   &__left
     display flex
