+++ client/apps/edit/components/header/test/index.test.js
@@ -1,9 +1,11 @@
 import Backbone from 'backbone'
 import React from 'react'
+import Article from '../../../../../models/article'
 import { EditHeader } from '../index'
 import { mount } from 'enzyme'
 import { Fixtures } from '@artsy/reaction-force/dist/Components/Publishing'
 import Icon from '@artsy/reaction-force/dist/Components/Icon'
+import colors from '@artsy/reaction-force/dist/Assets/Colors'
 
 describe('Edit Header Controls', () => {
   let props
@@ -17,7 +19,7 @@ describe('Edit Header Controls', () => {
         publishArticle: jest.fn(),
         saveArticle: jest.fn()
       },
-      article: new Backbone.Model(Fixtures.StandardArticle),
+      article: new Article(Fixtures.StandardArticle),
       channel: new Backbone.Model(),
       edit: {
         isSaved: true,
@@ -59,7 +61,7 @@ describe('Edit Header Controls', () => {
       const component = mount(
         <EditHeader {...props} />
       )
-      expect(component.find(Icon).first().props().color).toBe('limegreen')
+      expect(component.find(Icon).first().props().color).toBe(colors.greenRegular)
     })
 
     it('Content indicates non-completion', () => {
@@ -67,7 +69,7 @@ describe('Edit Header Controls', () => {
       const component = mount(
         <EditHeader {...props} />
       )
-      expect(component.find(Icon).first().props().color).toBe('#ccc')
+      expect(component.find(Icon).first().props().color).toBe(colors.grayMedium)
     })
 
     it('Display indicates completion if complete', () => {
@@ -75,7 +77,7 @@ describe('Edit Header Controls', () => {
       const component = mount(
         <EditHeader {...props} />
       )
-      expect(component.find(Icon).last().props().color).toBe('limegreen')
+      expect(component.find(Icon).last().props().color).toBe(colors.greenRegular)
     })
 
     it('Display indicates non-completion', () => {
@@ -83,7 +85,7 @@ describe('Edit Header Controls', () => {
       const component = mount(
         <EditHeader {...props} />
       )
-      expect(component.find(Icon).last().props().color).toBe('#ccc')
+      expect(component.find(Icon).last().props().color).toBe(colors.grayMedium)
     })
   })
 
@@ -243,7 +245,7 @@ describe('Edit Header Controls', () => {
       const component = mount(
         <EditHeader {...props} />
       )
-      expect(component.html()).toMatch('color: rgb(247, 98, 90)')
+      expect(component.find('button').at(4).props().style.color).toBe(colors.redMedium)
     })
 
     it('Color is green if isSaving', () => {
@@ -251,7 +253,7 @@ describe('Edit Header Controls', () => {
       const component = mount(
         <EditHeader {...props} />
       )
-      expect(component.find('button').at(4).props().style.color).toBe('limegreen')
+      expect(component.find('button').at(4).props().style.color).toBe(colors.greenRegular)
     })
 
     it('Color is black if isSaved', () => {
