+++ client/apps/edit/components/layout/components/yoast/yoast.styl
@@ -2,7 +2,7 @@
   position fixed
   width calc(100% \- 110px)
   top 55px
-  z-index 3
+  z-index 4
   .edit-seo
     &__header-container
       cursor pointer
