+++ client/apps/edit/index.styl
@@ -1,10 +1,8 @@
 @import './components/stylus_lib'
 @import './components/layout'
 @import './components/header'
-@import './components/header2'
 @import './components/display'
-@import './components/display2'
-@import './components/display2/preview'
+@import './components/display/preview'
 @import './components/admin'
 @import './components/mobile'
